+++ CHANGELOG.md
@@ -5,11 +5,12 @@
 - Widen supported Java version, produce Java 15 bytecode instead of 17
 - Improved error messages precision for invalid expected string blocks
 - Removed constraint on expected string block starting line, it can now be anywhere after the `expect(` call
+
 ## Bugfixes
 
-- promotion failure when promoting content in parallel within the same file. 
-Promotion is now synchronized on the ExpectTests class.
+- promotion failure when promoting content in parallel within the same file. Promotion is now synchronized on the ExpectTests class.
 - promotion failure when `expect(` was written within the expected string block on the same line as the actual `expect(` call.
+- promotion failure when using string interpolation within expected string block
 
 # 0.1.0
 Mostly bug fixes related to erroneous expect("...") calls without proper triple-quoted block formatting.